requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
nltk>=3.8.1
textblob>=0.17.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
sqlalchemy>=2.0.20
aiosqlite>=0.19.0
requests>=2.31.0 